+++ client/src/App.css
@@ -1,8 +1,3 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-
 .app {
   display: flex;
   flex-direction: column;
