+++ client/src/components/Home.css
@@ -1,8 +1,3 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-
 .about {
   max-width: 40%;
   display: block;
@@ -18,7 +13,7 @@
 }
 
 .appName {
-  font-family: "Baloo Paaji 2", cursive;
+  font-family: 'Baloo Paaji 2', cursive;
   font-weight: 800;
   font-size: 1.2em;
 }
@@ -40,7 +35,7 @@
 }
 
 .callToAction {
-  font-family: "Baloo Paaji 2", cursive;
+  font-family: 'Baloo Paaji 2', cursive;
   font-weight: 600;
   font-size: 1.2em;
   text-decoration: underline;
