+++ client/src/components/ItemDetail.css
@@ -67,8 +67,6 @@
   width: 100%;
 }
 
-
-
 @media (max-width: 1024px) {
   .name {
     font-size: 32px;
