+++ client/src/components/ItemDetail.jsx
@@ -1,23 +1,23 @@
-import React, { Component } from "react";
-import { Link, Redirect } from "react-router-dom";
+import React, { Component } from 'react';
+import { Link, Redirect } from 'react-router-dom';
 
-import Layout from "./shared/Layout";
+import Layout from './shared/Layout';
 
-import { getItem, deleteItem } from "../services/items";
-import "./ItemDetail.css";
-import mascot from "../Images/Mascot_over_red_2.png";
+import { getItem, deleteItem } from '../services/items';
+import './ItemDetail.css';
+import mascot from '../images/mascot_over_red.png';
 
 class ItemDetail extends Component {
   constructor() {
     super();
     this.state = {
       item: {
-        name: "",
-        description: "",
-        imgURL: "",
-        preferredQty: "",
-        onHandQty: "",
-        _id: "",
+        name: '',
+        description: '',
+        imgURL: '',
+        preferredQty: '',
+        onHandQty: '',
+        _id: '',
       },
       deleted: false,
     };
@@ -39,7 +39,7 @@ class ItemDetail extends Component {
     const { user } = this.props;
 
     if (deleted) {
-      return <Redirect to={"/items"} />;
+      return <Redirect to={'/items'} />;
     }
 
     return (
