+++ client/src/components/ItemEdit.css
@@ -137,10 +137,6 @@ button:hover {
     max-width: 100vw;
   }
 
-  .edit-item-edit {
-    /* margin-left: 1.5em; */
-  }
-
   .edit-item-image {
     width: 300px;
     margin: 15px;
