+++ client/src/components/Items.css
@@ -9,8 +9,8 @@
 
 .sort-label {
   font-size: 20px;
-  text-shadow: .5px .5px 2px white;
-  font-family: "Montserrat";
+  text-shadow: 0.5px 0.5px 2px white;
+  font-family: 'Montserrat';
   font-weight: 500;
   padding-right: 10px;
   color: black;
@@ -20,7 +20,7 @@
   cursor: pointer;
   border: none;
   font-size: 18px;
-  font-family: "Montserrat";
+  font-family: 'Montserrat';
   background-color: rgba(255, 255, 255, 0.808);
 }
