+++ client/src/components/Search.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import './Search.css';
 
 const Search = (props) => {
