+++ client/src/components/SignIn.css
@@ -1,5 +1,3 @@
-@import url("https://fonts.googleapis.com/css?family=Montserrat:200,400,500,700&display=swap");
-
 button {
   cursor: pointer;
   font-size: 1rem;
@@ -80,12 +78,10 @@ button {
   margin: 0.5em auto;
   font-size: 1rem;
   border-radius: 3px;
-  
+
   outline: transparent;
   transition: all 0.4s ease;
 }
-
-
 
 .form-container form input:valid {
   border: transparent;
@@ -211,7 +207,7 @@ button {
 }
 
 .item a::before {
-  content: "<-";
+  content: '<-';
 }
 
 .item-container .item {
