+++ client/src/components/SignUp.css
@@ -1,5 +1,3 @@
-@import url("https://fonts.googleapis.com/css?family=Montserrat:200,400,500,700&display=swap");
-
 button {
   cursor: pointer;
   font-size: 1rem;
@@ -115,7 +113,7 @@ button {
 .form-container form label {
   margin-top: 0.6em;
   text-shadow: 0.5px 0.25px 0.1px white;
-  font-family: "Montserrat";
+  font-family: 'Montserrat';
   font-weight: 400;
   text-align: left;
 }
@@ -226,7 +224,7 @@ button {
 }
 
 .item a::before {
-  content: "<-";
+  content: '<-';
 }
 
 .item-container .item {
