+++ client/src/components/shared/Footer.jsx
@@ -1,5 +1,6 @@
-import React from "react";
-import "./Footer.css";
+import React from 'react';
+
+import './Footer.css';
 
 function Footer() {
   return (
