+++ client/src/components/shared/Layout.css
@@ -1,9 +1,3 @@
-html,
-body {
-  margin: 0;
-  height: 100%;
-}
-
 .layout {
   display: flex;
   flex-direction: column;
