+++ client/src/images/mascot.png
