+++ client/src/index.css
@@ -1,9 +1,21 @@
+@import url('https://fonts.googleapis.com/css2?family=Baloo+Paaji+2:wght@600;800&display=swap');
+@import url('https://fonts.googleapis.com/css?family=Montserrat:200,400,500,700&display=swap');
+
+* {
+  margin: 0;
+  padding: 0;
+}
+
+html {
+  height: 100%;
+}
+
 body {
   margin: 0;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   padding: 0;
-  font-family: "Montserrat", sans-serif;
+  font-family: 'Montserrat', sans-serif;
   overflow: hidden;
   height: 100%;
   background-image: url(./Images/background.png);
@@ -14,7 +26,7 @@ body {
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
